[2주차] App 컴포넌트 분리
- index.jsx 에선 ReactDOM을 render 해주는 역할을 하고 있음
- 그래서 App 컴포넌트를 따로 분리함
- 분리 후 필요없어진 Page 컴포넌트는 삭제
